+++ noodles-bgzf/examples/bgzf_read.rs
@@ -2,14 +2,14 @@
 //!
 //! The result matches the output of `bgzip --decompress --stdout <src>`.
 
-use std::{env, fs::File, io};
+use std::{env, io};
 
 use noodles_bgzf as bgzf;
 
 fn main() -> io::Result<()> {
     let src = env::args().nth(1).expect("missing src");
 
-    let mut reader = File::open(src).map(bgzf::Reader::new)?;
+    let mut reader = bgzf::reader::Builder::default().build_from_path(src)?;
 
     let stdout = io::stdout();
     let mut writer = stdout.lock();